es loadtest

+++ scripts/es_pvc_resize.sh
@@ -16,45 +16,68 @@
 
 set -e
 
-# Set StatefulSet name
+
+# Set resize
+SS=elasticsearch-data
+NS=default
+VOLUME=data
+CAPACITY=128Gi
+HOOK_NAME=pre-stop-hook.sh
+
+#
 SS=jw5-pvc-debug
 NS=jwtest
-
-# Set post resize capacity for PVs
-CAPACITY=15Gi
+VOLUME=data
+CAPACITY=16Gi
+HOOK_NAME=bin.bash
 
 selector=`kubectl get statefulset --namespace=$NS $SS -o go-template='--selector={{range $k, $v := .spec.selector.matchLabels}}{{$k}}={{$v}},{{end}}' | sed 's/,$//'`
 pods=(`kubectl get pod --namespace=$NS $selector -o go-template='{{range .items}}{{.metadata.name}} {{end}}'`)
-pvcs=(`kubectl get pod --namespace=$NS $selector -o go-template='{{range .items}}{{range .spec.volumes}}{{if eq .name "data"}}{{.persistentVolumeClaim.claimName}} {{end}}{{end}}{{end}}'`)
+pvcs=(`kubectl get pod --namespace=$NS $selector -o go-template='{{range .items}}{{range .spec.volumes}}{{if eq .name "'$VOLUME'"}}{{.persistentVolumeClaim.claimName}} {{end}}{{end}}{{end}}'`)
 
+resized_pvcs=()
 for pvc in ${pvcs[@]}; do
     echo "patch pvc $pvc"
-    kubectl patch pvc --namespace=$NS $pvc --patch='{"spec":{"resources":{"requests":{"storage":"'$CAPACITY'"}}}}'
+    current_capacity=`kubectl get pvc --namespace=$NS $pvc -o go-template='{{.spec.resources.requests.storage}}'`
+    if [[ $current_capacity < $CAPACITY ]]; then
+        kubectl patch pvc --namespace=$NS $pvc --patch='{"spec":{"resources":{"requests":{"storage":"'$CAPACITY'"}}}}'
+	resized_pvcs+=($pvc)
+    else
+        echo "already applied"
+    fi
 done
+if [[ ${#resized_pvcs[@]} == 0 ]]; then
+    echo "no PVC to resize"
+    exit 1
+fi
 
 echo "scale down ss $SS"
 kubectl scale statefulset --namespace=$NS $SS --replicas=0
 
-echo ""
-while [[ ! -z `kubectl get pod --namespace=$NS $selector -o go-template='{{range .items}}{{.metadata.name}} {{end}}'` ]]; do
-    terminating=(`kubectl get pod --namespace=$NS $selector | awk '/Terminating/{print($1)}'`)
-    for pod in ${terminating[@]}; do
-        echo -ne "\e[0K\rterminating pod $pod"
-        pid=`kubectl exec --namespace=$NS $pod -- ps aux | awk '/pre-stop-hook.sh/{print($2)}'`
-        if [[ ! -z $pid ]]; then
-            echo -n ", killing hook with pid $pid"
-            kubectl exec --namespace=$NS $pod -- kill $pid
-            echo " - killed"
-        fi
-        sleep 2
+if [[ ! -z $HOOK_NAME ]]; then
+    echo ""
+    while [[ ! -z `kubectl get pod --namespace=$NS $selector -o go-template='{{range .items}}{{.metadata.name}} {{end}}'` ]]; do
+        terminating=(`kubectl get pod --namespace=$NS $selector | awk '/Terminating/{print($1)}'`)
+        for pod in ${terminating[@]}; do
+            echo -ne "\e[0K\rterminating pod $pod"
+            pid=`kubectl exec --namespace=$NS $pod -- ps aux | awk '/'$HOOK_NAME'/{print($2)}'`
+            if [[ ! -z $pid ]]; then
+                echo -n ", killing hook with pid $pid"
+                kubectl exec --namespace=$NS $pod -- kill $pid
+                echo " - killed"
+            else
+                echo -n ", hook probably already killed"
+            fi
+            sleep 2
+        done
     done
-done
+fi
 
 declare -A finished_pvc
-echo -n "finished_pvc:0  pvcs:${#pvcs[@]} $statuses"
-while [[ ${#finished_pvc[@]} != ${#pvcs[@]} ]]; do
+echo -n "finished_pvc:0  pvcs:${#resized_pvcs[@]} $statuses"
+while [[ ${#finished_pvc[@]} != ${#resized_pvcs[@]} ]]; do
     statuses=''
-    for pvc in ${pvcs[@]}; do
+    for pvc in ${resized_pvcs[@]}; do
         status=$(kubectl --namespace=$NS get pvc $pvc --output=go-template --template='{{range .status.conditions}}{{.type}} {{end}}')
         statuses="$statuses $pvc:[$status]"
         if [[ $status == *FileSystemResizePending* ]]; then
@@ -62,7 +85,7 @@ while [[ ${#finished_pvc[@]} != ${#pvcs[@]} ]]; do
         fi
         sleep 1
     done
-    echo -ne "\e[0K\rfinished_pvc:${#finished_pvc[@]} pvcs:${#pvcs[@]} - $statuses"
+    echo -ne "\e[0K\rfinished_pvc:${#finished_pvc[@]} pvcs:${#resized_pvcs[@]} - $statuses"
 done
 echo ""
 
@@ -78,4 +101,4 @@ while [[ ${#running_pods[@]} != ${#pods[@]} ]]; do
 done
 echo ""
 
-kubectl get pvc --namespace=$NS ${pvcs[@]}
+kubectl get pvc --namespace=$NS ${resized_pvcs[@]}
